Add shared data drawer for persistent right drawer, update MUI
