+++ process-flow-diagram-component/package.json
@@ -23,8 +23,8 @@
   "devDependencies": {
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
-    "@mui/icons-material": "^6.4.8",
-    "@mui/material": "^6.4.8",
+    "@mui/icons-material": "^7.2.0",
+    "@mui/material": "^7.2.0",
     "@types/node": "^20.12.7",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
